Extract saga files into separate own modules
