+++ src/app/sagas/index.js
@@ -1,6 +1,16 @@
-import pageSaga from './pageSaga';
+import pageCategoriesSaga from './pageCategoriesSaga';
+import pagesSaga from './pagesSaga';
+import pageSubCategoriesSaga from './pageSubCategoriesSaga';
+import productCategoriesSaga from './productCategoriesSaga';
+import productsSaga from './productsSaga';
+import productSubCategoriesSaga from './productSubCategoriesSaga';
 
 export default function* rootSaga() {
   yield []
-    .concat(pageSaga);
+    .concat(pageCategoriesSaga)
+    .concat(pagesSaga)
+    .concat(pageSubCategoriesSaga)
+    .concat(productCategoriesSaga)
+    .concat(productsSaga)
+    .concat(productSubCategoriesSaga);
 }
